pyproject.toml
[tool.poetry]
name = "fastpea"
version = "0.1.0"
description = ""
authors = ["iokiris"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
pymongo = "^4.6.1"
django-environ = "^0.11.2"
setuptools = "^69.0.2"
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
